scripts/auto_gen_utils/requirements.txt
asn1crypto==1.5.1
atomicwrites==1.4.1; python_version < '3.7.0'
atomicwrites==1.1.5; python_version >= '3.7.0'
attrs==18.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==7.1.2
configparser~=4.0.2
cryptography==3.2.1
defusedxml==0.6.0
dotmap==1.2.20
enum34==1.1.6
funcsigs==1.0.2
future==0.16.0
gitdb2==2.0.3
GitPython==2.1.9
idna==2.7
ipaddress==1.0.22
jira==1.0.15; python_version < '3.7.0'
jira==3.2.0; python_version >= '3.7.0'
more-itertools==4.3.0
oauthlib==2.1.0
ordereddict==1.1
parse==1.8.4
pathlib2==2.3.3
pbr==4.1.0
pluggy==0.7.1
pycparser==2.18
PyGithub<1.57
pyOpenSSL==18.0.0
pytest==3.7.2
pytz==2018.5
requests==2.25.1
requests-oauthlib==0.8.0; python_version < '3.7.0'
requests-oauthlib>=1.1.0; python_version >= '3.7.0'
requests-toolbelt==0.8.0
scandir==1.9.0
simplejson==3.16.0
six==1.11.0
smmap2==2.0.3
uritemplate==0.6
urllib3==1.26.2
verlib==0.1
voluptuous==0.11.5
flake8==3.7.9
stashy==0.5
Jinja2==2.11.2
python-dateutil==2.7.3
vcrpy==1.13.0
setuptools==44.1.1
packaging==19.0
recordclass==0.12.0.1
slack-sdk==3.17.0; python_version >= '3.7.0'

# Pretty old, but a lot of our scripts in TeamCity still run with Python 2,
# and all we need is Object Storage put_object
oci==2.104.3; python_version >= '3.7.0'
oci==2.53.0; python_version < '3'